move update_manifests_and_abi function to migration
